% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/agg_summaries.R
\name{agg_summaries}
\alias{agg_summaries}
\title{agg_summaries}
\usage{
agg_summaries(foo, var, group, FUN = mean, ...)
}
\description{
Aggregate summaries on foo
}
\examples{
\dontrun{
agg_summaries(dat, var = "latitude", group = "code", mean) \%>\% head
agg_summaries(dat, var = "latitude", group = "code", median) \%>\% head
agg_summaries(dat, var = "latitude", group = "code", quantile, .1) \%>\% head
}
}

