% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calculate_prevalence.R
\name{calculate_prevalence}
\alias{calculate_prevalence}
\title{Calculate prevalence within each region of aggregation}
\usage{
calculate_prevalence(data, pops = NULL, conf.level = 0.95,
  region.head = "region", scale = 1)
}
\arguments{
\item{data}{A line list-type data frame (one line per case), including a column with the region in which the case occurred}

\item{pops}{A data frame containing one line per region in "data", with first column for region name and second for population}

\item{conf.level}{Level of confidence interval required for prevalence estimates.}

\item{scale}{Scaling with which to report prevalence (per head, per 100 000, etc.)}

\item{region}{A string containing the name of the column specifying the region for each case}
}
\value{
A data frame containing total population, number of cases and prevalence by region
}
\description{
Calculate prevalence within each region of aggregation
}

