% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calculate_stratified_prevalence.R
\name{calculate_stratified_prevalence}
\alias{calculate_stratified_prevalence}
\title{Calculate prevalence within each region of aggregation}
\usage{
calculate_stratified_prevalence(data, stratify.vars, pops = NULL,
  region.head = "region", conf.level = 0.95, scale = 1)
}
\arguments{
\item{data}{A line list-type data frame (one line per case), including a column with the region in which the case occurred and columns with stratification variables.}

\item{stratify.vars}{A vector of strings containing the names of the column specifying the stratification variables. These columns should be factors.}

\item{pops}{A data frame containing one line per region/stratification variable combination in \code{data}. The first column should give the region name, then should come the stratification variables, and finally the population size in that region and stratum. If \code{pops} is \code{NULL}, the population in each area and stratum is taken as 1, and numbers of cases (rather than prevalences) are calculated.}

\item{region.head}{A string containing the name of the column specifying the region for each case.}

\item{conf.level}{Level of confidence interval required for prevalence estimates.}

\item{scale}{Scaling with which to report prevalence (per head, per 100 000, etc.)}
}
\value{
A list containing 1) \code{prevalence.list}, a list of data frames containing total population, number of cases and prevalence by region, for each unique combination of stratification variables. 2) \code{stratification.levels}, a vector of the stratification levels, in the same order that the stratified prevalence data frames are presented.
}
\description{
Calculate prevalence within each region of aggregation
}
\examples{
my.data <- data.frame(	county.id = ceiling(3*runif(10)),
                     	age = rlnorm(10),
							sex = factor(floor(2*runif(10)), levels=c(0,1), labels=c("male", "female"))
                     )

my.data$age.group = floor(my.data$age)
my.populations <- my.data[!duplicated(my.data[,c(1,3,4)]),]
my.populations$population <- nrow(my.data) + ceiling(abs(rnorm(nrow(my.populations), 10, 5)))

# example without any population data
calculate_stratified_prevalence(my.data, stratify.vars=c("sex", "age.group"), region.head="county.id")

# example with stratified population data
calculate_stratified_prevalence(my.data, stratify.vars=c("sex", "age.group"), pops=my.populations, region.head="county.id")
}

