% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/density_data.R
\name{density_data}
\alias{density_data}
\title{Density data}
\usage{
density_data(points, weights = NULL)
}
\arguments{
\item{points}{SpatialPointsDataFrame containing the line list of patients. Should include: 1) a slot @coords with an n x 2 vector giving the x and y coordinates of each case. (More than one case can have the same co-ordinates.) 2) a slot @bbox giving the maximum and minimum x and y co-ordinates.}

\item{weights}{If required, the local population density at each (x, y) coordinate.}
}
\value{
A data frame containing x and y co-ordinates, and (weighted) counts. For use by other functions including estimate_density.
}
\description{
Function description
process a SpatialPointsDataFrame to give numbers of cases at each unique (x,y) location. Weighting by 1/(local population density) gives corrected numbers which can be used to calculate estimated prevalence.
}
\details{
Process a SpatialPointsDataFrame to give numbers of cases at each unique (x,y) location. Weighting by 1/(local population density) gives corrected numbers which can be used to calculate estimated prevalence.
}

