% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/estimate_density.R
\name{estimate_density}
\alias{estimate_density}
\title{Estimate density}
\usage{
estimate_density(density_data)
}
\arguments{
\item{density_data}{A data frame with columns x, y and (corrected) number of individuals, as output by the density_data function.}
}
\value{
A SpatialGridDataFrame object containing intensity/prevalence estimates (@data$v), at a grid of points (@grid).
}
\description{
function to estimate 2d density, from data frame output by density_data
Produce a 2-d estimated intensity function (cases per unit area) from an (x, y, (corrected) counts) data frame.
}

