% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/check_data.R
\name{fix_dates}
\alias{fix_dates}
\title{Converts columns with dates into a R date class}
\usage{
fix_dates(df, date = NULL, format = NULL)
}
\arguments{
\item{df}{The input data.frame}

\item{date}{The column name containing the dates}

\item{format}{The format of the date.}
}
\description{
Using lubridate, this function will transform dates from mdy, mdy_h, mdy_hm, mdy_hms (or starting with day instead of month) into valid date classes. It will return an error it if cannot coerce the date itself.
}
\examples{
\dontrun{
new <- fix_dates(goat_data, "start.date", "dmy")
}
}

