% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{md_geocoder}
\alias{md_geocoder}
\title{Geocode a location}
\usage{
md_geocoder(location = NULL)
}
\arguments{
\item{location}{The name of a location}
}
\description{
Geocode a location
}

