% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/check_data.R
\name{test_lat}
\alias{test_lat}
\title{Valid latitude}
\usage{
test_lat(x)
}
\arguments{
\item{x}{A latitude}
}
\description{
Test for valid latitude. Between -90 and 90
}
\examples{
\dontrun{
expect_true(test_lat(dat$latitude))
}
}

