% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/unusual_prevalence.r
\name{unusual_performance_region}
\alias{unusual_performance_region}
\title{unusual.performance.region}
\usage{
unusual_performance_region(data, pops, region.head, region.i)
}
\arguments{
\item{data}{foo}

\item{pops}{foo}

\item{region.head}{foo}

\item{region.i}{foo}
}
\description{
more text describing the function
}

