% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calculate_prevalence.R
\name{calculate_prevalence}
\alias{calculate_prevalence}
\title{calculate_prevalence}
\usage{
calculate_prevalence(data, pops = NULL, conf.level = 0.95,
  region.head = "region", scale = 1)
}
\arguments{
\item{data}{A line list-type data frame (one line per case), including a column with the region in which the case occurred}

\item{pops}{A data frame containing one line per region in \code{data}, with first column for region name and second for population}

\item{conf.level}{Level of confidence interval required for prevalence estimates.}

\item{region.head}{A string containing the name of the column specifying the region for each case}

\item{scale}{Scaling with which to report prevalence (per head, per 100 000, etc.)}
}
\value{
A data frame containing total population, number of cases and prevalence by region
}
\examples{
my.data <- data.frame(	county.id = ceiling(3*runif(10)),
                     	age = rlnorm(10),
							sex = factor(floor(2*runif(10)), levels=c(0,1), labels=c("male", "female"))
                     )

my.populations <- data.frame(county.id = 1:3,
								population = c(10, 50, 100)
								)

# example without population of each region
calculate_prevalence(my.data, region.head="county.id")

# example with populations
calculate_prevalence(my.data, region.head="county.id", pops=my.populations)
}

