% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/unusual_prevalence.R
\name{calculate_prevalence_unusual_pval}
\alias{calculate_prevalence_unusual_pval}
\title{calculate_prevalence_unusual_pval}
\usage{
calculate_prevalence_unusual_pval(data, pops = NULL, conf.level = 0.95,
  region.head = "region", scale = 1)
}
\arguments{
\item{data}{placeholder text}

\item{pops}{placeholder text}

\item{conf.level}{placeholder text}

\item{region.head}{placeholder text}

\item{scale}{placeholder text}
}
\description{
more text here
}

